+++ src/coverage_array.py
@@ -1,4 +1,3 @@
-# Version 2.2.a
 
 import array
 import statistics as sts