+++ src/dependencies.py
@@ -1,4 +1,3 @@
-# Version 2.3.a
 
 import re
 import os