+++ src/output.py
@@ -1,4 +1,3 @@
-# Version 2.2.a
 
 import os
 import sys